\name{get.whas100.dataset}
\alias{get.whas100.dataset}
\title{Function that retrieves the WHAS100 dataset.}
\usage{
get.whas100.dataset()
}
\value{
A data.frame containing the WHAS100 dataset
}
\description{
This function retrieves the well known Worcester Heart Attack dataset with 100 individuals (WHAS100). This dataset is taken from the book by Hosmer D, Lemeshow S, May S. Applied Survival Analysis: Regression Modeling of Time to Event Data, 2nd edition. John Wiley and Sons Inc., New York, NY, 2008.
}
\examples{
whas100_data <- get.whas100.dataset()
}
